Fix stuff from review
