+++ crewai_tools/tools/__init__.py
@@ -22,16 +22,15 @@ from .firecrawl_scrape_website_tool.firecrawl_scrape_website_tool import (
 )
 from .firecrawl_search_tool.firecrawl_search_tool import FirecrawlSearchTool
 from .github_search_tool.github_search_tool import GithubSearchTool
-from .hyperbrowser_browser_use_tool.hyperbrowser_browser_use_tool import (
+from .hyperbrowser_tools import (
     HyperbrowserBrowserUseTool,
-)
-from .hyperbrowser_claude_computer_use.hyperbrowser_claude_computer_use import (
     HyperbrowserClaudeComputerUseTool,
+    HyperbrowserOpenAICuaTool,
+    HyperbrowserCrawlTool,
+    HyperbrowserExtractTool,
+    HyperbrowserLoadTool,
+    HyperbrowserScrapeTool,
 )
-from .hyperbrowser_crawl_tool.hyperbrowser_crawl_tool import HyperbrowserCrawlTool
-from .hyperbrowser_extract_tool.hyperbrowser_extract_tool import HyperbrowserExtractTool
-from .hyperbrowser_load_tool.hyperbrowser_load_tool import HyperbrowserLoadTool
-from .hyperbrowser_scrape_tool.hyperbrowser_scrape_tool import HyperbrowserScrapeTool
 from .json_search_tool.json_search_tool import JSONSearchTool
 from .linkup.linkup_search_tool import LinkupSearchTool
 from .llamaindex_tool.llamaindex_tool import LlamaIndexTool
