+++ crewai_tools/tools/hyperbrowser_tools/__init__.py
@@ -1,0 +1,23 @@
+from .hyperbrowser_scrape_tool.hyperbrowser_scrape_tool import HyperbrowserScrapeTool
+from .hyperbrowser_crawl_tool.hyperbrowser_crawl_tool import HyperbrowserCrawlTool
+from .hyperbrowser_extract_tool.hyperbrowser_extract_tool import HyperbrowserExtractTool
+from .hyperbrowser_load_tool.hyperbrowser_load_tool import HyperbrowserLoadTool
+from .hyperbrowser_claude_computer_use.hyperbrowser_claude_computer_use import (
+    HyperbrowserClaudeComputerUseTool,
+)
+from .hyperbrowser_browser_use_tool.hyperbrowser_browser_use_tool import (
+    HyperbrowserBrowserUseTool,
+)
+from .hyperbrowser_openai_cua_tool.hyperbrowser_openai_cua_tool import (
+    HyperbrowserOpenAICuaTool,
+)
+
+__all__ = [
+    "HyperbrowserBrowserUseTool",
+    "HyperbrowserClaudeComputerUseTool",
+    "HyperbrowserCrawlTool",
+    "HyperbrowserExtractTool",
+    "HyperbrowserLoadTool",
+    "HyperbrowserOpenAICuaTool",
+    "HyperbrowserScrapeTool",
+]
