+++ crewai_tools/tools/hyperbrowser_tools/common/errors.py
@@ -1,0 +1,1 @@
+NOT_INITIALIZED_ERROR = RuntimeError("Hyperbrowser not properly initialized")
